First level refactoring of Program.cs
